# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yap API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/reports/cdr": {
            "get": {
                "tags": [
                    "CDR"
                ],
                "summary": "Get Call Detail Records report",
                "operationId": "212743d161bbe68ebb7c267be8cab799",
                "parameters": [
                    {
                        "name": "service_body_id",
                        "in": "query",
                        "description": "ID of the service body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_range_start",
                        "in": "query",
                        "description": "Start date for the report range",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_range_end",
                        "in": "query",
                        "description": "End date for the report range",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "recurse",
                        "in": "query",
                        "description": "Whether to include records from child service bodies",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Call Detail Records retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "call_sid": {
                                                "type": "string"
                                            },
                                            "from": {
                                                "type": "string"
                                            },
                                            "to": {
                                                "type": "string"
                                            },
                                            "start_time": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "end_time": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "duration": {
                                                "type": "integer"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "service_body_id": {
                                                "type": "integer"
                                            },
                                            "volunteer_id": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "volunteer_name": {
                                                "type": "string",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/events/status": {
            "get": {
                "tags": [
                    "EventStatus"
                ],
                "summary": "Get all event statuses",
                "operationId": "6c5f582c0ee947c4877a303fece7250f",
                "responses": {
                    "200": {
                        "description": "List of event statuses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "callsid": {
                                                "type": "string"
                                            },
                                            "event_id": {
                                                "type": "integer"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "EventStatus"
                ],
                "summary": "Create a new event status",
                "operationId": "410743153504c29a76ad087c9d6e5ea8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "callsid",
                                    "event_id",
                                    "status"
                                ],
                                "properties": {
                                    "callsid": {
                                        "description": "Call SID",
                                        "type": "string"
                                    },
                                    "event_id": {
                                        "description": "Event ID",
                                        "type": "integer"
                                    },
                                    "status": {
                                        "description": "Status of the event",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event status created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "callsid": {
                                            "type": "string"
                                        },
                                        "event_id": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to create event status"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "BearerAuth": []
        }
    ]
}